// fe/src/types/about.ts
export type AboutCardProps = {
  img: string;
  title: string;
  description?: string;
  address?: string;
  isPrizeSection?: boolean;
  firstPlace?: string;
  secondPlace?: string;
  thirdPlace?: string;
  lastPlace?: string;
};

export type AboutGameRoundsProps = {
  img: string;
  title: string;
  description: string;
};